swiftshop-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
razorpay
httpx
